+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Current main branch.
     - `Token` matches arbitrary tokens
 - Values
   - Turned Value::Int to crate [num-bigint](https://crates.io/crates/num-bigint), replaced Value::Addr by the same type.
+  - Definition of mutable objects; Imutable objects push a clone of, mutable objects push a ref on the object.
 - Builtins
   - Added `dict_push()`
   - Renamed `dict_update()` into `dict_merge()`